bips instead of pips

+++ contracts/interfaces/IUniswapV3Pair.sol
@@ -26,7 +26,7 @@ interface IUniswapV3Pair {
     function price1CumulativeLast() external view returns (uint256);
 
     // derived state
-    function getFee() external view returns (uint24 fee);
+    function getFee() external view returns (uint16 fee);
     function getVirtualSupply() external view returns (uint112 virtualSupply);
     function getCumulativePrices() external view returns (uint256 price0Cumulative, uint256 price1Cumulative);
 }
